fix: win + rpm and 5.0 check

+++ src/init/wazuh/wazuh.sh
@@ -192,12 +192,14 @@ WazuhUpgrade()
         rm -f /etc/ossec-init.conf
     fi
 
-    # Remove old databases
-    if [ -f $PREINSTALLEDDIR/queue/syscollector/db/local.db ]; then
-        rm -f $PREINSTALLEDDIR/queue/syscollector/db/local.db
-    fi
-    if [ -f $PREINSTALLEDDIR/queue/fim/db/fim.db ]; then
-        rm -f $PREINSTALLEDDIR/queue/fim/db/fim.db
+    # Remove old databases if upgrading from 4.X
+    if [ $MAJOR = 4 ]; then
+        if [ -f $PREINSTALLEDDIR/queue/syscollector/db/local.db ]; then
+            rm -f $PREINSTALLEDDIR/queue/syscollector/db/local.db
+        fi
+        if [ -f $PREINSTALLEDDIR/queue/fim/db/fim.db ]; then
+            rm -f $PREINSTALLEDDIR/queue/fim/db/fim.db
+        fi
     fi
 
     # Replace and delete ossec group along with ossec users
